requirements.txt
llama-index
llama-hub
langchain
openai
python-dotenv
chromadb
beautifulsoup4
lxml
tqdm
youtube_transcript_api
numpy
transformers
sentencepiece
huggingface-hub
plotly
nbformat
python-dateutil
chromadb
torch
spacy

# vllm
# google-generativeai